+++ src/app/message_commands.rs
@@ -327,7 +327,7 @@ fn message_update_event(message: MessageInfo) -> AppEvent {
             fields: MessageUpdateEventFields {
                 poll: message.poll,
                 content: message.content,
-                sticker_names: Some(message.sticker_names),
+                stickers: Some(message.stickers),
                 mentions: Some(message.mentions),
                 mention_everyone: Some(message.mention_everyone),
                 mention_roles: Some(message.mention_roles),