+++ src/discord/ids.rs
@@ -117,5 +117,8 @@ pub mod marker {
     pub struct RoleMarker;
 
     #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
+    pub struct StickerMarker;
+
+    #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
     pub struct UserMarker;
 }